py_backend/requirements.txt
google-cloud-secret-manager==2.17.0
openai==1.3.8
text-generation==0.6.0
Pillow==9.5.0
firebase_admin==6.3.0
PyYAML==6.0.1
numpy==1.24.2
opencv-python-headless==4.7.0.72
pydub==0.25.1
tiktoken==0.3.3
librosa==0.10.0.post2
# for matching the assets from llm descriptions
fuzzywuzzy==0.18.0
# address log: 'Using slow pure-python SequenceMatcher. Install python-Levenshtein to remove this warning'
python-Levenshtein==0.23.0

# Downgrade urllib3 to fix current bug in the dependency chain firebase-admin -> requests -> urllib3.
# See link below and upgrade version when patch is out + test with https://github.com/ParamaxAi/py_backend/issues/84.
# https://stackoverflow.com/questions/76175361/firebase-authentication-httpresponse-object-has-no-attribute-strict-status
# -> You'd get a 401 error in the browser when going from the first to the second workflow step (+ a nice CORS error).
urllib3==1.26.15

# server
Flask==2.3.2
flask-cors==3.0.10
gunicorn==20.1.0

# CLIP inference: install transformers, see Dockerfile for CPU-only pytorch
transformers==4.30.2
